optimize script
